+++ training/benchmarks/yolov5/pytorch/model/__init__.py
@@ -7,6 +7,7 @@ from experimental import attempt_load
 from utils.general import check_suffix,intersect_dicts
 from utils.torch_utils import torch_distributed_zero_first
 from utils.downloads import attempt_download, is_url
+import yaml
 
 LOCAL_RANK = int(os.getenv('LOCAL_RANK', -1))  # https://pytorch.org/docs/stable/elastic/run.html
 RANK = int(os.getenv('RANK', -1))
@@ -15,26 +16,29 @@ WORLD_SIZE = int(os.getenv('WORLD_SIZE', 1))
 
 def create_model(config):
     # model = torchvision.models.mobilenet_v2()
-    cfg = config.cfg
+    weights, cfg, resume = config.weights, config.cfg, config.resume
+
     # hpy = "path/to/hpy.yaml"
-    hyp = config.hyp
-    resume = config.resume
+    # hyp = config.hyp
+    # Hyperparameters
+    if isinstance(config.hyp, str):
+        with open(config.hyp, errors='ignore') as f:
+            hyp = yaml.safe_load(f)  # load hyps dict
+    
     device = config.device
-    nc = 1
+    # nc = 1 if single_cls else int(data_dict['nc'])  # number of classes
+    # coco number of class: 80
+    nc = 80
     # Model
-    weights = config.weights
     check_suffix(weights, '.pt')  # check weights
     pretrained = weights.endswith('.pt')
     if pretrained:
-        with torch_distributed_zero_first(LOCAL_RANK):
-            weights = attempt_download(weights)  # download if not found locally
         ckpt = torch.load(weights, map_location='cpu')  # load checkpoint to CPU to avoid CUDA memory leak
         model = Model(cfg or ckpt['model'].yaml, ch=3, nc=nc, anchors=hyp.get('anchors')).to(device)  # create
         exclude = ['anchor'] if (cfg or hyp.get('anchors')) and not resume else []  # exclude keys
         csd = ckpt['model'].float().state_dict()  # checkpoint state_dict as FP32
         csd = intersect_dicts(csd, model.state_dict(), exclude=exclude)  # intersect
         model.load_state_dict(csd, strict=False)  # load
-        # LOGGER.info(f'Transferred {len(csd)}/{len(model.state_dict())} items from {weights}')  # report
     else:
         model = Model(cfg, ch=3, nc=nc, anchors=hyp.get('anchors')).to(device)  # create
     return model
